requirements.txt
accelerate==0.30.1
peft==0.11.1
bitsandbytes==0.43.1
transformers==4.41.2
diffusers==0.29.0
torch  
datasets
tqdm
Pillow
torchvision
